Move cast_sign_loss to its own module
